KMCH-main/requirements.txt
flask 
flask-cors 
pyaudio 
numpy 
scipy 
pandas 
google-generativeai
flask
 flask-cors 
python-dotenv
 pymongo
  google-generativeai
   pyaudio


flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
python-dotenv==1.0.0
google-generativeai==0.3.1
pymongo==4.6.1
deepgram-sdk==3.2.0
pyaudio==0.2.14